--- Merging r71224 into '.': A    test/Transforms/ScalarRepl/2009-05-08-I1Crash.ll U    lib/Transforms/Scalar/ScalarReplAggregates.cpp
fix RewriteStoreUserOfWholeAlloca to use the correct type size
method, fixing a crash on PR4146.  While the store will
ultimately overwrite the "padded size" number of bits in memory,
the stored value may be a subset of this size.  This function
only wants to handle the case where all bits are stored.

llvm-svn=71229