We currently only run ipsccp at LTO time, which is silly.  It subsumes ipconstprop and doesn't take much time.  Just run it in its place.
This adds a testcase for it, which I plan to expand to cover other
"integration" cases, where we expect the optimizer to be able to
eliminate various things.  Due to phase order issues we've regressed
in a number of areas and integration tests are the only way I see to
prevent this.

llvm-svn=85729