When we are deserializing the lexical decls of a DeclContext from PCH, notify the PCHReader to hold off passing Decls to the consumer until the DeclContext is fully prepared.
Before, due to recursive loading, we could be in a situation where we would try to deserialize the decls of a DeclContext which was already doing that, and bad things would happen. In the specific case I encountered, the lexical decls would
form a cycle and we would enter infinite loop territory.

llvm-svn=109857