Allow GUIs that use the LLDB framework to translate "control + char" into strings that can be written into the file handle that will feed the debugger input file.
Most GUIs will create a PTY where the master end is written into and the slave end is given to the debugger as input, output and error. Since pseudo terminals don't have height and width and dont handle CTRL+D and other control characters the same way, we need a way to emulate this. 

llvm-svn=199931