Merging r242604: ------------------------------------------------------------------------ r242604 | chandlerc | 2015-07-17 20:14:02 -0700 (Fri, 17 Jul 2015) | 12 lines
[cmake] Support enabling -Werror in the OpenMP runtime CMake build and
clean up the build.

This disables all of the Clang warnings that fire for me when building
libomp.so on Linux with a recent Clang binary. Lots of these should
probably be fixed, but I want to at least get the build warning-clean
and make it easy to keep that way.

I also switched a bunch of the warnings that are used both for C and C++
compiles to check the flag with C compilation test.

Differential Revision: http://reviews.llvm.org/D11253
------------------------------------------------------------------------

llvm-svn=242909