Merging r324722: ------------------------------------------------------------------------ r324722 | labath | 2018-02-09 10:40:03 +0100 (Fri, 09 Feb 2018) | 17 lines
llgs-test: Parse and store register info recieved from lldb-server

Summary:
Right now the test client is not parsing register values correctly,
which is manifesting itself in one test failing on 32-bit architectures
(pr36013). This parses the information from the qRegisterInfo packets
and stores it in the client, which will enable fixing the parsing in a
follow up commit.

I am also adding a new templated SendMessage overload, which enables one
to send a message get a parsed response in a single call.

Reviewers: eugene, davide

Subscribers: lldb-commits

Differential Revision: https://reviews.llvm.org/D43076
------------------------------------------------------------------------

llvm-svn=325562